requirements.txt
langchain==0.1.0
openai==1.3.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
chromadb==0.4.18
pyyaml==6.0.1
python-dotenv==1.0.0
tiktoken==0.5.2